Update draft-westerlund-tsvwg-dtls-over-sctp-bis.md
The length numbers given for DTLS is not valid for DTLS 1.2 or DTLS 1.3. I don't think we need to go into the details of input plaintext size, plaintext size with added content type, compressions and protections overhead, and record header. They are different in all versions of DTLS.

Either we set the supported maximum length of SCTP user messages to something large enough for DTLS like 2^15 or we put something even bigger like 1024 kB which would be enough for current 3GPP use.