Resumption #76
Resumption #76

Some more information and considerations regarding resumption and tickets. There was a lot of questions from developers regarding ticket and resumption in EAP-TLS 1.3 so it might be good to give some information.

This is not critical. It is just informational information. Please review. Could also be in some other part of the document.